atomic-design - restructure components
