+++ src/components/atoms/Spinner.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import { spinnerStyle } from '../style';
+import { spinnerStyle } from '../../style';
 
 const animationDelay = (delay) => ({
   MozAnimationDelay: delay,
