+++ src/components/atoms/index.js
@@ -1,0 +1,2 @@
+export { Spinner } from './Spinner';
+export { TextInput } from './TextInput';
