+++ src/components/container/Card.js
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types'; // eslint-disable-line no-unused-vars
 import React, { Component } from 'react';
-import { cardStyle } from '../style';
+import { cardStyle } from '../../style';
 
 /**
  * Card component with a light shadow.
