+++ src/components/molecules/index.js
@@ -1,0 +1,6 @@
+export { Button } from './Button';
+export { ComboBoxItem } from './ComboBoxItem';
+export { Day } from './Day';
+export { DisabledDay } from './DisabledDay';
+export { SelectItem } from './SelectItem';
+export { Separator } from './Separator';
