+++ src/components/organisms/index.js
@@ -1,0 +1,4 @@
+export { ComboBox } from './ComboBox';
+export { Rating } from './Rating';
+export { Select } from './Select';
+export { Toggle } from './Toggle';
