+++ src/components/templates/index.js
@@ -1,0 +1,1 @@
+export { DatePicker } from './DatePicker';
