+++ src/index.js
@@ -1,16 +1,24 @@
-import { Button } from './components/Button';
-import { Card } from './components/Card';
-import { Choice } from './components/Choice';
-import { ComboBox } from './components/ComboBox';
-import { DatePicker } from './components/DatePicker';
-import { Option } from './components/Option';
-import { Placeholder } from './components/Placeholder';
-import { Rating } from './components/Rating';
-import { Select } from './components/Select';
-import { Separator } from './components/Separator';
-import { Spinner } from './components/Spinner';
-import { TextInput } from './components/TextInput';
-import { Toggle } from './components/Toggle';
+import {
+  Spinner,
+  TextInput,
+} from './components/atoms';
+import {
+  Card,
+  Choice,
+  Option,
+  Placeholder,
+} from './components/container';
+import {
+  Button,
+  Separator,
+} from './components/molecules';
+import {
+  ComboBox,
+  Rating,
+  Select,
+  Toggle,
+} from './components/organisms';
+import { DatePicker } from './components/templates';
 
 import {
   buttonConfig,
